feat(quotation&order): created a autofill flow when converting quotation to orders
This commit has some bugs that needs to be work on and it is not the final build.
It includes some backend fixes for quotation.

And this commit also includes some migration fixes.

+++ database/migrations/2025_12_09_124630_create_employee_data_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('employee_details')) {
+            return;
+        }
+
         Schema::create('employee_details', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
@@ -36,6 +40,6 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('user_address');
+        Schema::dropIfExists('employee_details');
     }
 };
